Part 1: Extract low-level test framework into mentat_db::debug for re-use.
